Easy API
1) Moved handhskaing as a part of internal WebView API

2) Placed Invalidation right after backing store changes
i think this is the best place to invalidate- now paints without having to switch workspaces
