+++ Tools/MiniBrowser/haiku/BrowserApp.h
@@ -21,12 +21,8 @@ public:
 	void MessageReceived(BMessage *message);
 	void ReadyToRun();
 private:
-	void LocalMessage(BMessage*);
-	void GlobalMessage(BMessage*);
 	BWebView* webView;
 	status_t result;
-	map<string,BLooper*> looperMapping;
-	map<string,BMessage*> messengerMapping;
 	BrowserWindow* fWindow;
 };
